#### Reference Issue
This PR begins the work of #1130, extending librosa to support multi-channel wherever possible.


#### What does this implement/fix? Explain your changes.

This branch will serve as the target for smaller pull requests implementing multi-channel support throughout the package.

As a first step, I've implemented multi-channel stft.  It's numerically equivalent to our previous implementation, and comparably efficient for single-channel inputs.
For multi-channel inputs, it's not as efficient as I think is possible, but optimization will take a bit of sleuthing.


#### Any other comments?

This PR will not be merged until #1130 is complete.