@bmcfee 2021-07-07
# Progress tracker

## beat

    - [ ] `tempo` : yes
    - [ ] `plp` : yes

## decompose

    - [ ] `decompose` : maybe, could require some clever reshaping. not obvious how to do it.
    - [ ] `hpss` : yes
    - [ ] `nn_filter` : probably not, sparse matrix support is a dealbreaker

## effects

    - [ ] `hpss` : yes
    - [ ] `harmonic` : yes
    - [ ] `percussive` : yes
    - [ ] `time_stretch` : yes
    - [ ] `pitch_shift` : yes
    - [ ] `remix` : yes-ish, zc alignment could be tricky
    - [ ] `preemphasis`: yes
    - [ ] `deemphasis`: yes

## filters
    - [ ] `diagonal_filter`: could require some thought

## onset

    - [ ] `onset_strength`: yes
    - [ ] `onset_strength_multi`: yes
    - [ ] `onset_backtrack`: probably not

## segment
    - [ ] `cross_similarity` : should be possible, except for sparsity
    - [ ] `recurrence_matrix` : should be possible, except for sparsity
    - [ ] `recurrence_to_lag` : should be possible, except for sparsity
    - [ ] `lag_to_recurrence` : should be possible, except for sparsity
    - [ ] `timelag_filter` : maybe
    - [ ] `path_enhance`: maybe, depends on `filters.diagonal_filter`


## sequence

    - [ ] `dtw` : should be possible, could get ugly
    - [ ] `rqa` : should be possible, could get ugly
    - [ ] `viterbi` : yes
    - [ ] `viterbi_discriminative` : yes
    - [ ] `viterbi_binary` : yes
---

## core
    - [ ] `stream`: n/a, I think?
    - [ ] `to_mono`: yes
    - [ ] `resample`: yes, n/a
    - [ ] `get_duration`: yes, n/a
    - [ ] `autocorrelate`: yes
    - [ ] `lpc`: yes, could get ugly
    - [ ] `zero_crossings`: yes

    - [ ] `cqt`: yes, sparsity will be a problem
    - [ ] `hybrid_cqt`: yes
    - [ ] `pseudo_cqt`: yes
    - [ ] `icqt`: yes
    - [ ] `griffinlim_cqt`: yes
    - [ ] `vqt`: yes

    - [ ] `salience`: yes
    - [ ] `interp_harmonics`: yes
    - [ ] `harmonics_1d`: yes
    - [ ] `harmonics_2d`: could be possible, but not sure if it makes sense
    
    - [ ] `estimate_tuning`: yes
    - [ ] `pitch_tuning`: yes
    - [ ] `piptrack`: yes

## feature

    - [ ] `chroma_stft`: yes
    - [ ] `chroma_cqt`: yes 
    - [ ] `chroma_cens`: yes
    - [ ] `melspectrogram`: yes
    - [ ] `mfcc`:yes 
    - [ ] `rms`: yes
    - [ ] `spectral_centroid`: yes
    - [ ] `spectral_bandwidth`: yes
    - [ ] `spectral_flatness`: yes
    - [ ] `spectral_rolloff`: yes
    - [ ] `poly_features` : yes
    - [ ] `tonnetz` : yes
    - [ ] `zero_crossing_rate` : yes
    - [ ] `tempogram` : yes
    - [ ] `fourier_tempogram` : yes
    - [ ] `delta` : yes
    - [ ] `stack_memory` : yes
    - [ ] `mel_to_stft`: yes, will need nnls
    - [ ] `mel_to_audio` : yes
    - [ ] `mfcc_to_mel` :yes
    - [ ] `mfcc_to_audio`: yes

    - [ ] `yin`: yes
    - [ ] `pyin`: yes

    - [x] `stft`: yes
    - [ ] `istft`: yes
    - [ ] `magphase` : yes
    - [ ] `iirt`: yes
    - [ ] `reassigned_spectrogram`: yes
    - [ ] `phase_vocoder`: yes
    - [ ] `fmt`: yes
    - [ ] `pcen`: yes
    - [ ] `griffinlim`: yes


## util

    - [x] `frame`: yes
    - [ ] `pad_center`: yes
    - [ ] `fix_length`: yes
    - [ ] `fix_frames`: yes
    - [ ] `index_to_slice`: maybe
    - [ ] `softmask`: yes
    - [ ] `stack`: yes
    - [ ] `sync`: yes-ish, could be tricky

    - [ ] `axis_sort`: maybe, not clear what it would mean
    - [ ] `normalize`: yes
    - [ ] `shear`: maybe; dense only?

    - [ ] `localmax`: yes
    - [ ] `localmin`: yes
    - [ ] `nnls`: maybe
    - [ ] `valid_audio`: yes