@bmcfee 2021-07-07
I've done a first cut through of basic functionality, and updated the checklist with functions that already supported multi-channel intrinsically.  At this point, we have working stft, istft, griffinlim, loading, streaming, resampling, and a handful of utility functions needed to make those work.

The next major hurdle in the core module is CQT.  This is theoretically easy because it relies on STFT, but our use of sparse basis projection presents a bit of a problem as mentioned [here](https://github.com/librosa/librosa/issues/1130#issuecomment-844017923).  In short, we represent the DFT of the CQT basis as a sparse (CSR) array `fft_basis`, and compute `fft_basis.dot(D)` where `D` is the STFT of the input.  

https://github.com/librosa/librosa/blob/d9b72af964820fc796425c55d3dccd7f6b908bf9/librosa/core/constantq.py#L513-L514

Sparse matrices can only matrix multiply against 1d or 2d arrays (vectors or matrices), and the more general einstein summation / tensor product is not supported.  This means that we won't be able to get away with just swapping out a dot product for an einsum and having everything work gracefully.

A few solutions:

1. Use basis sparsification when the input is mono, and fall back to dense multiplication for multi-channel.  This can be done after sparsification, something like `if D.ndim > 2: fft_basis = fft_basis.toarray(), np.einsum(fft_basis, D, ...) ; else: fft_basis.dot(D)`.  This will give numerically identical results, but ruin our efficiency for multichannel.
2. Iterate over channels, and dot product per channel.  This is simple and will work, but will also limit us to flat multichannel arrays (eg, `y=(n_channels, n_samples)`).  We would not be able to support higher order objects directly (`y=(n_sensors, n_channels, n_samples)`, for example).
3. Write our own projection routine with numba.  CSR matrices are easy enough to parse and work with under the hood (we do it in other places), so this shouldn't be too hard.  It will be a bit more work, but it might end up being the most flexible option.
4. Not support sparsification for multichannel inputs.  I don't like this option, as it's more restrictive than (1), but it is worth mentioning because it would at least save us one round-trip through sparse matrix construction.