@bmcfee 2021-07-08
Flagged a few more functions that already work without modification.  `interp_harmonics` seems to work, but I'm not sure the dimension manipulation is what we want:

```python
In [27]: D.shape
Out[27]: (2, 1025, 3949)

In [26]: librosa.interp_harmonics(np.abs(D), librosa.fft_frequencies(), [1,2,3],
    ...:  axis=1).shape
Out[26]: (3, 2, 1025, 3949)
```
whereas I'd probably expect the output to have shape `(2, 3, 1025, 3949)`.