@lostanlen 2021-07-08
Thank you @bmcfee for starting this endeavor!
I appreciate the 4-option summary. In my opinion:

1. breaks the principle of minimum surprise and might even create antipatterns (users writing an ad hoc implementation of (2) in their own code)
2. looks good to me. i don't understand why this approach is restricted to 2-D inputs though? can't we `np.view` the input tensor as having a shape `(-1, frequency, time)`  for the sake of iteration? and then run another `np.view` on the result so that it gets reshaped at N-D? (with N>=3)
3. that approach seems ideal. i don't know how difficult it is to write numba code, even less when making no assumptions on the dimension of of the inputs. 
4. same remark as (1)



> I'd probably expect the output to have shape (2, 3, 1025, 3949).

me too