@bmcfee 2021-08-31
Now that viterbi is multi-channel aware, here's a revised implementation of pyin

<details>
<summary> Multi-channel pyin implementation</summary>

```python
def pyin(
    y,
    fmin,
    fmax,
    sr=22050,
    frame_length=2048,
    win_length=None,
    hop_length=None,
    n_thresholds=100,
    beta_parameters=(2, 18),
    boltzmann_parameter=2,
    resolution=0.1,
    max_transition_rate=35.92,
    switch_prob=0.01,
    no_trough_prob=0.01,
    fill_na=np.nan,
    center=True,
    pad_mode="reflect",
):
   
    if fmin is None or fmax is None:
        raise ParameterError('both "fmin" and "fmax" must be provided')

    # Set the default window length if it is not already specified.
    if win_length is None:
        win_length = frame_length // 2

    if win_length >= frame_length:
        raise ParameterError(
            "win_length={} cannot exceed given frame_length={}".format(
                win_length, frame_length
            )
        )

    # Set the default hop if it is not already specified.
    if hop_length is None:
        hop_length = frame_length // 4

    # Check that audio is valid.
    #util.valid_audio(y, mono=False)

    # Pad the time series so that frames are centered
    if center:
        padding = [(0, 0) for _ in y.shape]
        padding[-1] = (frame_length //2, frame_length //2)
        y = np.pad(y, padding, mode=pad_mode)
        

    # Frame audio.
    y_frames = util.frame(y, frame_length=frame_length, hop_length=hop_length)

    # Calculate minimum and maximum periods
    min_period = max(int(np.floor(sr / fmax)), 1)
    max_period = min(int(np.ceil(sr / fmin)), frame_length - win_length - 1)

    # Calculate cumulative mean normalized difference function.
    yin_frames = _cumulative_mean_normalized_difference(
        y_frames, frame_length, win_length, min_period, max_period
    )

    # Parabolic interpolation.
    parabolic_shifts = _parabolic_interpolation(yin_frames)
    
    # Find Yin candidates and probabilities.
    # The implementation here follows the official pYIN software which
    # differs from the method described in the paper.
    # 1. Define the prior over the thresholds.
    thresholds = np.linspace(0, 1, n_thresholds + 1)
    beta_cdf = scipy.stats.beta.cdf(thresholds, beta_parameters[0], beta_parameters[1])
    beta_probs = np.diff(beta_cdf)

    n_bins_per_semitone = int(np.ceil(1.0 / resolution))
    n_pitch_bins = int(np.floor(12 * n_bins_per_semitone * np.log2(fmax / fmin))) + 1

    def _helper(a, b):
        return __pyin_helper(a, b, thresholds, boltzmann_parameter, 
                                 beta_probs, no_trough_prob, min_period, 
                                 fmin, n_pitch_bins, n_bins_per_semitone)
            
    helper = np.vectorize(_helper, signature='(f,t),(k,t)->(1,d,t),(j,t)')
    
    # Construct transition matrix.
    max_semitones_per_frame = round(max_transition_rate * 12 * hop_length / sr)
    transition_width = max_semitones_per_frame * n_bins_per_semitone + 1
    # Construct the within voicing transition probabilities
    transition = sequence.transition_local(
        n_pitch_bins, transition_width, window="triangle", wrap=False
    )
    # Include across voicing transition probabilities
    transition = np.block(
        [
            [(1 - switch_prob) * transition, switch_prob * transition],
            [switch_prob * transition, (1 - switch_prob) * transition],
        ]
    )
    
    p_init = np.zeros(2 * n_pitch_bins)
    p_init[n_pitch_bins:] = 1 / n_pitch_bins
    
    observation_probs, voiced_prob = helper(yin_frames, parabolic_shifts)
    states = sequence.viterbi(observation_probs, transition, p_init=p_init)
    
    # Find f0 corresponding to each decoded pitch bin.
    freqs = fmin * 2 ** (np.arange(n_pitch_bins) / (12 * n_bins_per_semitone))
    f0 = freqs[states % n_pitch_bins]
    voiced_flag = states < n_pitch_bins
    
    if fill_na is not None:
        f0[~voiced_flag] = fill_na
    
    return f0[..., 0, :], voiced_flag[..., 0, :], voiced_prob[..., 0, :]


def __pyin_helper(yin_frames, parabolic_shifts, thresholds, 
                      boltzmann_parameter, beta_probs, no_trough_prob, min_period, 
                      fmin, n_pitch_bins, n_bins_per_semitone):

    yin_probs = np.zeros_like(yin_frames)
    for i, yin_frame in enumerate(yin_frames.T):
        # 2. For each frame find the troughs.
        is_trough = util.localmin(yin_frame, axis=0)
        is_trough[0] = yin_frame[0] < yin_frame[1]
        (trough_index,) = np.nonzero(is_trough)

        if len(trough_index) == 0:
            continue

        # 3. Find the troughs below each threshold.
        trough_heights = yin_frame[trough_index]
        trough_thresholds = trough_heights[:, None] < thresholds[None, 1:]

        # 4. Define the prior over the troughs.
        # Smaller periods are weighted more.
        trough_positions = np.cumsum(trough_thresholds, axis=0) - 1
        n_troughs = np.count_nonzero(trough_thresholds, axis=0)
        trough_prior = scipy.stats.boltzmann.pmf(
            trough_positions, boltzmann_parameter, n_troughs
        )
        trough_prior[~trough_thresholds] = 0

        # 5. For each threshold add probability to global minimum if no trough is below threshold,
        # else add probability to each trough below threshold biased by prior.
        probs = np.sum(trough_prior * beta_probs, axis=1)
        global_min = np.argmin(trough_heights)
        n_thresholds_below_min = np.count_nonzero(~trough_thresholds[global_min, :])
        probs[global_min] += no_trough_prob * np.sum(
            beta_probs[:n_thresholds_below_min]
        )

        yin_probs[trough_index, i] = probs

    yin_period, frame_index = np.nonzero(yin_probs)

    # Refine peak by parabolic interpolation.
    period_candidates = min_period + yin_period
    period_candidates = period_candidates + parabolic_shifts[yin_period, frame_index]
    f0_candidates = sr / period_candidates
    
    # Find pitch bin corresponding to each f0 candidate.
    bin_index = 12 * n_bins_per_semitone * np.log2(f0_candidates / fmin)
    bin_index = np.clip(np.round(bin_index), 0, n_pitch_bins).astype(int)

    # Observation probabilities.
    observation_probs = np.zeros((2 * n_pitch_bins, yin_frames.shape[1]))
    observation_probs[bin_index, frame_index] = yin_probs[yin_period, frame_index]
    
    voiced_prob = np.clip(np.sum(observation_probs[:n_pitch_bins, :],
                                      axis=0, keepdims=True), 
                             0, 1)
    observation_probs[n_pitch_bins:, :] = (1 - voiced_prob) / n_pitch_bins

    return observation_probs[np.newaxis], voiced_prob
```

</details>

It's a little rough, but it works and is already a bit faster than our current implementation:

```python
# Current single-channel code
>>> %%timeit 
... librosa.pyin(y[0], 30, 600, fill_na=-1)
... librosa.pyin(y[1], 30, 600, fill_na=-1)
3.31 s ± 131 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# New version on single channels independently
>>> %%timeit
... pyin(y[0], 30, 600, fill_na=-1)
... pyin(y[1], 30, 600, fill_na=-1)
2.82 s ± 429 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# New version on both channels jointly
>>> %timeit pyin(y, 30, 600, fill_na=-1)
2.16 s ± 70.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
The proposed implementation first vectorizes to generate the observation likelihoods (inputs to viterbi) for each channel, collects the results, and then (implicitly) vectorizes again within viterbi to get the outputs.  This way, we cut down on redundancies in constructing the viterbi setup (transition matrices, etc), which for pyin can actually be fairly costly (states often number in the thousands).

The remaining bottleneck here is in the pyin frame loop.  I think we can't entirely squash this with numba because of the call out to `scipy.stats.boltzmann.pmf`, but there are a few parts of the loop that we might be able to factor out and vectorize.  I'll experiment with this before pushing changes.